clean up sub_menu_builder()
created functions in the LinuxFileConverterMenuProvider class to create sub menus and to append items to them. Argument structure for a few functions changed to (self, arguments), where arguments is a dict of arguments passed to the add_sub_menu_item() function. This change is needed to be able to pass different amounts of arguments to different kinds of functions while using the same function to construct the buttot which executes the specified callback