Change name to linux-file-converter-addon
nemo_action should be updated to dynamically have the name of the script file in it. Newly installed instances of linux-file-converter-addon should automatically be saved as linux-file-converter-addon.py and nautilus-file-converter-addon.nemo_action instead of the old name (nautlilus-fileconverter). This is hopefully backwards compatible for people having installed the program before this commit.